pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ppt-lint"
version = "0.1.0"
description = "Configuration-driven PowerPoint format checker and auto-fixer"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "python-pptx>=0.6.23",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "anthropic>=0.39.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
ppt-lint = "cli.ppt_lint:cli"

[tool.setuptools.packages.find]
include = ["cli*", "internal*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
